Current version of suckless.vim does not respect the orientation of splits in vim
The order of split commands in vim matters:
`split | vsplit` will cause the following scheme:
```
+--+--+
|  |  |
+--+--+
|     |
+-----+
```

while `vsplit | split` will create the following splitting:
```
+--+--+
|  |  |
+--+  +
|  |  |
+-----+
```

When suckless.vim is used it creates the second type scheme regardless
of the order.

In this merge I've replaced 'wincmd n' with 'wincmd a:direction' which
fixes the issue.